test(e2e): split swap/wrap/errors
